Make Updatables take a long instead of an int when update(...) is called.

+++ java/believe/app/Game.kt
@@ -12,15 +12,16 @@ import javax.inject.Inject
 /** The base class for a game.  */
 @Reusable
 class Game @Inject constructor(
-    @ApplicationTitle
-    private val applicationTitle: String, private val gameStateRunner: GameStateRunner,
+    @ApplicationTitle private val applicationTitle: String,
+    private val gameStateRunner: GameStateRunner,
     private val stateController: Lazy<StateController>,
     private val stateSelector: StateSelector
 ) : Game {
 
     override fun init(container: GameContainer) = stateSelector.selectState(stateController.get())
 
-    override fun update(container: GameContainer, delta: Int) = gameStateRunner.update(delta)
+    override fun update(container: GameContainer, delta: Int) =
+        gameStateRunner.update(delta.toLong())
 
     override fun render(container: GameContainer, g: org.newdawn.slick.Graphics) =
         gameStateRunner.render(Graphics(g))
